Implemented the new comment method and did some formatting to make them stand out

+++ public/js/delete.js
@@ -1,17 +1,15 @@
-const newDeleteHandler = async (evnt) => {
-  evnt.preventDefault();
-  console.log('You hit the button');
-  await fetch('/api/posts/', {
-    method: 'DELETE',
-    body: JSON.stringify({
-      id,
-      title,
-      content,
-    }),
-    headers: { 'Content-Type': 'application/json' },
-  });
+const delButtonHandler = async (event) => {
+  if (event.target.hasAttribute('data-id')) {
+    const id = event.target.getAttribute('data-id');
+    await fetch(`/api/posts/${id}`, {
+      method: 'DELETE',
+      headers: { 'Content-Type': 'application/json' },
+    });
+
+    document.location.replace('/dashboard');
+  }
 };
 
 document
   .querySelector('#delete-post-button')
-  .addEventListener('click', newDeleteHandler);
+  .addEventListener('click', delButtonHandler);
